Angular code for #382
Used $resource for fast and scalable development
